% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblind_discrete.R
\docType{data}
\name{colorblind}
\alias{colorblind}
\title{Color-blind colors
Colorblind-friendly colors for discrete values ####}
\format{
An object of class \code{character} of length 15.
}
\usage{
data('colorblind')
export
}
\description{
Color-blind colors
Colorblind-friendly colors for discrete values ####
}
\keyword{datasets}
