% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthodrome.R
\name{orthodrome}
\alias{orthodrome}
\title{Orthodromic distance}
\usage{
orthodrome(a, b)
}
\arguments{
\item{a}{lon, lat coordinate of point 1}

\item{b}{lon, lat coordinate of point 2}
}
\value{
distance in km
}
\description{
The great-circle distance or orthodromic distance is the
shortest distance between two points on the surface of a sphere,
measured along the surface of the sphere (as opposed to a straight line
through the sphere's interior)
}
\examples{
berlin <- c(52.517, 13.4)
tokyo <- c(35.7, 139.767)
orthodrome(berlin, tokyo)
}
