% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geochron.R
\name{read_geochron}
\alias{read_geochron}
\title{Read data from Geochron database}
\usage{
read_geochron(path, method = "U-Pb")
}
\arguments{
\item{path}{Path to the xls file.}

\item{method}{Character string defining the input type of the dataset. One of
"Ar-Ar", "Fission Track", "(U-Th)/He", and "U-Pb"}
}
\value{
list
}
\description{
Converts Excel spreadsheets from a Geochon download into a meta
data and isotope data table. \code{meta} contains all the sample information.
\code{isotopes} contains all the information of the measurements (i.e. isotopes, ages)
}
